# [quorum store][mempool] optimize pulling txns for when many txns are in progress