@bchocho 2023-11-01
Recent ad-hoc test: https://github.com/aptos-labs/aptos-core/actions/runs/6725257657/attempts/1#summary-18279247367
vs forge stable: https://github.com/aptos-labs/aptos-core/actions/runs/6716280366/attempts/1#summary-18257624474