@zekun000 2023-11-02
also I don't remember why this select requires a biased, if we don't have strong reasons we should default to use the randomized select