@bchocho 2023-11-02
I vaguely remember in a previous version, the backpressure was on/off and so having it absolutely up-to-date was a concern. However, with the current backpressure it's probably less of a concern. I'll remove biased and run some experiments