@bchocho 2023-11-02
Things look fine without biased. Here is a run with gauge instrumented (from #10772), it has at max 1 item in the queue: https://github.com/aptos-labs/aptos-core/actions/runs/6736434894