@bchocho 2023-11-03
For BTreeMap, it's hard to know if it's safe to remove a transaction from exclude_transactions_sorted when a batch is removed from batches_in_progress. I'm pretty sure there are edge cases when there could be transactions in multiple batches_in_progress, so we couldn't remove _all_ transactions within the removed batch. I suppose we could do some kind of reference counting or check the batch id before removal, but that makes it somewhat complex.

With the current approach, exclude_transactions_sorted is computed from scratch, so it's simpler (but possibly more inefficient)